# docs/swagger.yaml
basePath: /swagger/ebcp-service
definitions:
...
  api.StaticControlExhibitionRoomRequest:
    properties:
      command:
        type: string
      device_ip:
        type: string
      device_port:
        type: integer
    type: object
...
  common.PageGeneric-model_Ebcp_exhibition_room:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Ebcp_exhibition_room'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Ebcp_exhibition_room:
    properties:
      created_by:
        description: created_by
        type: string
      created_time:
        description: created_time
        type: string
      exhibition_hall_id:
        description: 所属展馆ID
        type: string
      exhibition_id:
        description: 所属展览ID
        type: string
      floor:
        description: 楼层(B1,F1,F2,F3...)
        type: string
      id:
        description: id
        type: string
      location:
        description: 展厅位置(西侧，西北侧)
        type: string
      name:
        description: 展厅名称
        type: string
      remarks:
        description: 备注
        type: string
      status:
        description: '状态（1: 正常, 2: 未使用, 3: 维修）'
        type: integer
      updated_by:
        description: updated_by
        type: string
      updated_time:
        description: updated_time
        type: string
    type: object
...
info:
  contact: {}
  description: ebcp-service API
  title: ebcp-service API
  version: "1.0"
paths:
...
  /ebcp-exhibition-room:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: location
        in: query
        name: location
        type: string
      - description: exhibition_hall_id
        in: query
        name: exhibition_hall_id
        type: string
      - description: floor
        in: query
        name: floor
        type: string
      - description: exhibition_id
        in: query
        name: exhibition_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: remarks
        in: query
        name: remarks
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Ebcp_exhibition_room'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - 展厅
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Ebcp_exhibition_room'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ebcp_exhibition_room'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - 展厅
...
  /ebcp-exhibition-room/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ebcp_exhibition_room'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - 展厅
  /ebcp-exhibition-room/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - 展厅
  /ebcp-exhibition-room/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - 展厅
  /ebcp-exhibition-room/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: _select
        in: query
        name: _select
        required: true
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: location
        in: query
        name: location
        type: string
      - description: exhibition_hall_id
        in: query
        name: exhibition_hall_id
        type: string
      - description: floor
        in: query
        name: floor
        type: string
      - description: exhibition_id
        in: query
        name: exhibition_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: remarks
        in: query
        name: remarks
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.PageGeneric-model_Ebcp_exhibition_room'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - 展厅
  /ebcp-exhibition-room/static-control:
    post:
      description: Static control an exhibition room by ID
      parameters:
      - description: Static Control Exhibition Room Request
        in: body
        name: command_request
        required: true
        schema:
          $ref: '#/definitions/api.StaticControlExhibitionRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: Static control exhibition room
      tags:
      - 展厅
...
swagger: "2.0"